requirements.txt
# Web框架
Flask>=3.0.0
Werkzeug>=3.0.0

# PDF转图片功能
pdf2image>=1.16.0
Pillow>=10.0.0